refactored voyager verifier into separate crate
