+++ crates/sozo/ops/src/migration_ui.rs
@@ -2,8 +2,9 @@
 
 use std::fmt;
 
+use sozo_voyager::VerificationUi;
 use spinoff::spinners::SpinnerFrames;
-use spinoff::{spinner, spinners, Spinner};
+use spinoff::{Spinner, spinner, spinners};
 
 /// A simple UI for the migration that can be used to display a spinner.
 pub struct MigrationUi {
@@ -91,5 +92,11 @@ impl MigrationUi {
         }
 
         self.spinner = Spinner::new(self.default_frames.clone(), text, None);
+    }
+}
+
+impl VerificationUi for MigrationUi {
+    fn update_text_boxed(&mut self, text: String) {
+        self.update_text_boxed(text);
     }
 }
